+++ packages/template-retail-react-app/app/components/product-view/index.jsx
@@ -36,8 +36,8 @@ import {STORE_LOCATOR_IS_ENABLED} from '@salesforce/retail-react-app/app/constan
 import {getConfig} from '@salesforce/pwa-kit-runtime/utils/ssr-config'
 import {useShopperBasketsV2Mutation as useShopperBasketsMutation} from '@salesforce/commerce-sdk-react'
 import {
-    useSFPaymentsEnabled,
-    useSFPayments
+    useSFPayments,
+    useExpressCheckoutEnabled
 } from '@salesforce/retail-react-app/app/hooks/use-sf-payments'
 
 // project components
@@ -217,7 +217,7 @@ const ProductView = forwardRef(
         const [pickupEnabled, setPickupEnabled] = useState(false)
         const storeName = selectedStore?.name
         const inventoryId = selectedStore?.inventoryId
-        const sfPaymentsEnabled = useSFPaymentsEnabled()
+        const {pdp: showExpressOnPDP} = useExpressCheckoutEnabled()
 
         const {disableButton, customInventoryMessage} = useMemo(() => {
             let shouldDisableButton = showInventoryMessage
@@ -470,7 +470,7 @@ const ProductView = forwardRef(
             }
 
             if (
-                sfPaymentsEnabled &&
+                showExpressOnPDP &&
                 !isProductASet &&
                 !isProductPartOfBundle &&
                 activeCurrency &&